Add pointer examples

+++ src/lessons/5-pointers/11-void.c
@@ -3,7 +3,7 @@
 int main() {
   int x = 5;
   void *ptr = &x;            // Cast int* to void*
-  double *int_ptr = (double*)ptr; // Cast void* back to int*
-  printf("Value of x: %g\n", *int_ptr);
+  int *int_ptr = (int *)ptr; // Cast void* back to int*
+  printf("Value of x: %d\n", *int_ptr);
   return 0;
 }
